Struct interface::Modifiers

#[repr(C)]

pub struct Modifiers {
    pub highway: c_double,
    pub rating_sum: c_double,
    pub rating_count: size_t,
    pub water: size_t,
    pub park: size_t,
}

Modifiers of an edge.

Every edge (road) has a few properties: whether it is a highway or a footpath, whether it is in a park, or near water, etc. This structure contains all that data in a slightly compressed format.

Fields
highway: c_double

Highway modifier, between 1.0 (footpath) and infinity (motorway)

rating_sum: c_double

Sum of all the ratings, or rating_count * average_rating

rating_count: size_t

Number of ratings, or rating_sum / average_rating

water: size_t

1 if the road is near water, else 0

park: size_t

1 if the road is in a park, else 0