Crate interface

Interface for use in the FFI.

This crate contains most of the necessary functions that are exported to the FFI interface.

Structs

Configuration
The actual configuration.

ConnIdVal
Iterator over all connections from a single node, with ID and Data.

DijkstraIterator
DijkstraIterator.

DijkstraNode
Structure yielded by a Dijkstra's iterator.

Edge
Road struct

Graph
Representation of a graph

IdVal
Representation of an edge and its ID, as yielded by a ConnId iterator.

IterConnIds
Iterator over all connections from a single node, with only ID.

IterEdges
Iterator over all connections from a single node, with only Data.

ListIds
Iterator over all node IDs in the graph.

Modifier_Dot
Modifier of the configuration.

Modifiers
Modifiers of an edge.

Node
Representation of a node.

Poison
Representation of a poisoned graph

Root
Iterator over a path generated by Dijkstra's algorithm.

Functions

graph_conn_idval_delete⚠
Delete the iterator

graph_conn_idval_new⚠
Retrieve an iterator over all edges and connected ids.

graph_conn_idval_next⚠
Advance the iterator

graph_connids_delete⚠
Delete the iterator

graph_connids_new⚠
Retrieve an iterator over all connection ids

graph_connids_next⚠
Advance the iterator

graph_contains⚠
Check if the graph contains a node with the given index

graph_delete⚠
Delete the graph, freeing its memory.

graph_dijkstra_choose⚠
Pick a random element from the dijkstraiterator that is at a decent distance. Rooting keeps working.

graph_dijkstra_delete⚠
Delete the iterator

graph_dijkstra_filter⚠
Filter the dijkstraiterator, so that it only returns nodes included in the node list. Rooting keeps working.

graph_dijkstra_next⚠
Advance the iterator

graph_dijkstra_root⚠
Root the iterator

graph_edges_delete⚠
Delete the iterator

graph_edges_new⚠
Retrieve an iterator over all edges

graph_edges_next⚠
Advance the iterator

graph_get⚠
Retrieve a node from the graph

graph_graph_generate⚠
Generate a collection of shortest paths from a starting point

graph_listids_delete⚠
Delete the iterator

graph_listids_new⚠
Retrieve an iterator over all node ids in the graph

graph_listids_next⚠
Advance the iterator

graph_new⚠
Create a new graph.

graph_poison⚠
Retrieve a new poisoned graph.

graph_poison_delete⚠
Delete the iterator

graph_poison_generate⚠
Generate a collection of shortest paths from a starting point

graph_root_delete⚠
Delete the iterator

graph_root_next⚠
Advance the iterator

graph_update_rating⚠
Update the rating of the road in the graph